feat: Add undo toast when a card is deleted

+++ src/components/Card/Card.tsx
@@ -1,6 +1,7 @@
 import React, { useState, useRef, useMemo } from 'react';
 import { observer } from 'mobx-react-lite';
 import { toast } from 'react-toastify';
+import cn from 'classnames';
 
 import { cards as cardsActions } from '../../core/actions';
 import CardDocument from '../../documents/card.doc';
@@ -12,6 +13,7 @@ import CardBadges from '../CardBadges/CardBadges';
 import DraggableCard from './DraggableCard';
 import MarkdownText from '../shared/MarkdownText';
 import { Card as CardModel } from '../../documents/card.doc';
+import { ToastUndo } from '../shared';
 import { useBoardsStore } from '../../store/boards';
 import { useSession } from '../providers/SessionProvider';
 
@@ -31,6 +33,8 @@ const Card = ({ card, index, listId, draggable = true }: CardProps) => {
   const { userDoc } = useSession();
   const toggleCheckbox = useMarkdownCheckbox(card.data.text);
 
+  const isHiddenRef = useRef(false);
+  const [, forceRenderer] = useState(false);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const cardRef = useRef(null);
   const [cardRect] = useRect(cardRef);
@@ -50,7 +54,9 @@ const Card = ({ card, index, listId, draggable = true }: CardProps) => {
     updateCard({ text: newText });
   };
 
-  const deleteCard = async () => {
+  const remove = async () => {
+    if (!isHiddenRef.current) return;
+
     const textBackup = card.data.text;
 
     await card.ref.delete();
@@ -63,7 +69,18 @@ const Card = ({ card, index, listId, draggable = true }: CardProps) => {
         title: card.data.title || '',
       },
     });
-    toast('Card was removed');
+  };
+
+  const undo = () => {
+    isHiddenRef.current = false;
+    forceRenderer(s => !s);
+  };
+
+  const deleteCard = async () => {
+    isHiddenRef.current = true;
+    toast(<ToastUndo title='Card removed' id={card.id} undo={undo} />, {
+      onClose: remove,
+    });
   };
 
   const handleClick = event => {
@@ -97,7 +114,10 @@ const Card = ({ card, index, listId, draggable = true }: CardProps) => {
       <>
         <div
           ref={cardRef}
-          className='card relative text-sm select-none cursor-pointer rounded my-2 mx-0 text-gray-900 break-words'
+          className={cn(
+            'card relative text-sm select-none cursor-pointer rounded my-2 mx-0 text-gray-900 break-words',
+            isHiddenRef.current && 'hidden'
+          )}
           style={{ backgroundColor: card.data.color, minHeight: '60px' }}
           onClick={handleClick}
           onKeyDown={handleKeyDown}
